Move flutter_js.dart to file_generators subdir.
Fix imports.
