+++ packages/flutter_tools/lib/src/isolated/devfs_web.dart
@@ -41,7 +41,7 @@ import '../vmservice.dart';
 import '../web/bootstrap.dart';
 import '../web/chrome.dart';
 import '../web/compile.dart';
-import '../web/flutter_js.dart' as flutter_js;
+import '../web/file_generators/flutter_js.dart' as flutter_js;
 import '../web/memory_fs.dart';
 
 typedef DwdsLauncher = Future<Dwds> Function({
