+++ packages/flutter_tools/lib/src/web/file_generators/flutter_js.dart
