+++ packages/flutter_tools/test/general.shard/build_system/targets/web_test.dart
@@ -13,7 +13,7 @@ import 'package:flutter_tools/src/build_system/build_system.dart';
 import 'package:flutter_tools/src/build_system/depfile.dart';
 import 'package:flutter_tools/src/build_system/targets/web.dart';
 import 'package:flutter_tools/src/globals.dart' as globals;
-import 'package:flutter_tools/src/web/flutter_js.dart' as flutter_js;
+import 'package:flutter_tools/src/web/file_generators/flutter_js.dart' as flutter_js;
 
 import '../../../src/common.dart';
 import '../../../src/fake_process_manager.dart';
